+++ src/SelectorFilter.js
